fix (redis_cache): update redis config to support any redis installation (#662)
* fix (redis_cache): update redis config to support any redis installation

* fix (redis_cache): not necessary try

* fix (redis_cache): fix secrets redis host

* fix (redis_cache): fix redis specs

+++ spec/rails_helper.rb
@@ -117,13 +117,9 @@ RSpec.configure do |config|
   end
   
   config.before(:suite) { FactoryBot.lint } if ENV["LINT"]
-  
-  config.before(:each) do
-    $reports.redis.flushdb
-  end
 
   config.after(:each) do
-    $reports.redis.flushdb
+    $reports.redis.flushdb if $reports.present?
   end
 
   config.expect_with :rspec do |expectations|
